improve conversions to float64
Improve the accuracy of converting a decimal value into a float64. These
conversions now shift from the decimal exponent to a binary exponent
using integer, instead of floating point, operations. This will likely
be slightly slower for values with large exponents, but allows the
calculation to use values with a greater range than the floating point
operations provide.